Fixes #87 - Will require clean rebuild of project in AS to take effect.
