+++ app/src/androidTest/java/com/raygun/raygun4android/sample/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.raygun.raygun4androidsample.sample
+package com.raygun.raygun4android.sample
 
 import android.support.test.InstrumentationRegistry
 import android.support.test.runner.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getTargetContext()
-        assertEquals("com.raygun.raygun4androidsample.sample", appContext.packageName)
+        assertEquals("com.raygun.raygun4android.sample", appContext.packageName)
     }
 }
