+++ app/src/main/java/com/raygun/raygun4android/sample/SecondActivity.kt
@@ -1,32 +1,25 @@
-package com.raygun.raygun4androidsample.sample
+package com.raygun.raygun4android.sample
 
-import android.app.Application
-import android.support.v7.app.AppCompatActivity
+import android.content.Context
+import android.content.Intent
 import android.os.Bundle
+import android.support.v7.app.AppCompatActivity
 import android.util.Log
-import com.raygun.raygun4android.RaygunClient
-import com.raygun.raygun4android.messages.shared.RaygunUserInfo
 import android.widget.Button
 import android.widget.TextView
+import com.raygun.raygun4android.RaygunClient
+import com.raygun.raygun4android.messages.crashreporting.RaygunBreadcrumbLevel
+import com.raygun.raygun4android.messages.crashreporting.RaygunBreadcrumbMessage
+import com.raygun.raygun4android.messages.shared.RaygunUserInfo
+import java.util.*
 
-class MainActivity : AppCompatActivity() {
+class SecondActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(R.layout.activity_second)
 
-        val initialCustomData = HashMap<String,String>()
-        initialCustomData["firstkey"] = "firstvalue"
-
-        // This sets up the client with the API key as provided in your AndroidManifest.xml
-        RaygunClient.init(applicationContext as Application)
-        RaygunClient.enableCrashReporting()
-        RaygunClient.enableRUM(this)
-        RaygunClient.shouldProcessBreadcrumbLocation(true)
-
-        RaygunClient.setCustomData(initialCustomData)
-
-        RaygunClient.setOnBeforeSend(SampleOnBeforeSend())
+        supportActionBar?.setDisplayHomeAsUpEnabled(true)
 
         val buttonSend = findViewById<Button>(R.id.button_send)
         val buttonCrash = findViewById<Button>(R.id.button_crash)
@@ -36,14 +29,13 @@ class MainActivity : AppCompatActivity() {
         val buttonSetUserB = findViewById<Button>(R.id.button_set_user_B)
         val textViewAppVersion = findViewById<TextView>(R.id.textView_appVersion)
         val textViewProviderVersion = findViewById<TextView>(R.id.textView_providerVersion)
-        val buttonSecondActivity = findViewById<Button>(R.id.button_secondActivity)
 
         buttonSend.setOnClickListener {
             val tw = HashMap<String,String>()
             tw["secondkey"] = "secondvalue"
 
             // Manual exception creation & sending
-            RaygunClient.send(Exception("Congratulations, you have sent errors with Raygun4Android"), null, tw)
+            RaygunClient.send(Exception("Congratulations, you have sent errors with Raygun4Android from SecondActivity"), null, tw)
         }
 
         buttonCrash.setOnClickListener {
@@ -74,7 +66,6 @@ class MainActivity : AppCompatActivity() {
             user.firstName = "User C"
             user.email = "e@f.com.com"
             RaygunClient.setUser(user)
-            RaygunClient.recordBreadcrumb("I'm now user C")
         }
 
         buttonSetUserB.setOnClickListener{
@@ -83,18 +74,29 @@ class MainActivity : AppCompatActivity() {
             user.firstName = "User D"
             user.email = "g@h.com"
             RaygunClient.setUser(user)
-            RaygunClient.recordBreadcrumb("I'm now user D")
         }
-
-        buttonSecondActivity.setOnClickListener {
-            startActivity(SecondActivity.getIntent(this@MainActivity))
-        }
-
 
         textViewAppVersion.text = "App ${BuildConfig.VERSION_NAME} (${BuildConfig.VERSION_CODE} ${BuildConfig.BUILD_TYPE})"
         textViewProviderVersion.text = "Provider ${com.raygun.raygun4android.BuildConfig.VERSION_NAME} (${com.raygun.raygun4android.BuildConfig.VERSION_CODE} ${BuildConfig.BUILD_TYPE})"
 
-        RaygunClient.recordBreadcrumb("I'm here in Main Activity")
+        RaygunClient.clearBreadcrumbs()
 
+        val customData = WeakHashMap<String,Any>()
+        customData["someKey"] = "someValue"
+        customData["someotherkey"] = "someothervalue"
+
+        val breadcrumbMessage = RaygunBreadcrumbMessage.Builder("I'm here in SecondActivity")
+            .level(RaygunBreadcrumbLevel.ERROR)
+            .category("Launch")
+            .lineNumber(78)
+            .methodName("onCreate")
+            .customData(customData)
+            .build()
+
+        RaygunClient.recordBreadcrumb(breadcrumbMessage)
+    }
+
+    companion object {
+        fun getIntent(context: Context): Intent = Intent(context, SecondActivity::class.java)
     }
 }
