+++ app/src/test/java/com/raygun/raygun4android/sample/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.raygun.raygun4androidsample.sample
+package com.raygun.raygun4android.sample
 
 import org.junit.Test
